Alter table CartLine

+++ src/main/resources/db/migration/V1_04__Create_Product_Table.sql
@@ -1,11 +1,11 @@
 CREATE TABLE IF NOT EXISTS product (
-                         id BIGINT PRIMARY KEY AUTO_INCREMENT,
+                         id INT PRIMARY KEY AUTO_INCREMENT,
                          title VARCHAR(255),
                          description TEXT,
                          main_picture VARCHAR(255),
                          status VARCHAR(50),
                          create_at DATETIME,
                          updated_at DATETIME,
-                         category_id BIGINT,
+                         category_id INT,
                          FOREIGN KEY (category_id) REFERENCES category (id)
 );
