+++ src/main/resources/db/migration/V1_06__Create_Image_Table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS image (
-                       id BIGINT PRIMARY KEY AUTO_INCREMENT,
+                       id INT PRIMARY KEY AUTO_INCREMENT,
                        img_path VARCHAR(255),
-                       variant_id BIGINT,
+                       variant_id INT,
                        FOREIGN KEY (variant_id) REFERENCES variant (id)
 );
