+++ src/main/resources/db/migration/V1_07__Create_OptionTable_Table.sql
@@ -1,4 +1,4 @@
 CREATE TABLE IF NOT EXISTS option_table (
-                              id BIGINT AUTO_INCREMENT PRIMARY KEY,
+                              id INT AUTO_INCREMENT PRIMARY KEY,
                               name VARCHAR(255)
 );
