+++ src/main/resources/db/migration/V1_08__Create_OptionValue_Table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS optionValue (
-                             id BIGINT AUTO_INCREMENT PRIMARY KEY,
+                             id INT AUTO_INCREMENT PRIMARY KEY,
                              value VARCHAR(255),
-                             option_id BIGINT,
+                             option_id INT,
                              FOREIGN KEY (option_id) REFERENCES option_table(id)
 );
