+++ src/main/resources/db/migration/V1_10__Create_Video_Table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS video (
-                       id BIGINT PRIMARY KEY AUTO_INCREMENT,
+                       id INT PRIMARY KEY AUTO_INCREMENT,
                        video_path VARCHAR(255),
-                       variant_id BIGINT,
+                       variant_id INT,
                        FOREIGN KEY (variant_id) REFERENCES variant (id)
 );
